+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 - The ``--data-generation-unique`` CLI option that forces Schemathesis to generate unique test cases only.
   This feature is also available as a hook in ``schemathesis.contrib.unique_data``.
 - ``schemathesis.checks.register`` as a new way to register checks.
+- ``schemathesis.targets.register`` as a new way to register targets.
 
 **Changed**
 
@@ -21,6 +22,7 @@ Changelog
 **Deprecated**
 
 - ``schemathesis.register_check`` in favor of ``schemathesis.checks.register``.
+- ``schemathesis.register_target`` in favor of ``schemathesis.targets.register``.
 
 **Fixed**